% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myconstr.R
\name{myconstr}
\alias{myconstr}
\title{This is my constructor function}
\usage{
myconstr(x, y)
}
\arguments{
\item{x}{vector of independent variable}

\item{y}{vector of dependent variable}
}
\value{
list of data and lm objects
}
\description{
This function will create a list output
}
\details{
The constructor makes an object filled with valuable material
Need to be more specific here.
}
\examples{
myconstr(x,y)
}
