% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{mynewf}
\alias{mynewf}
\title{Structures of the Data Set}
\usage{
mynewf(x, y, Var.equal = True, paired = True)
}
\arguments{
\item{x}{vector of data}

\item{y}{vector of data}

\item{paired}{}
}
\value{

}
\description{
Provides structure and attributes of data inputted.
}
\details{
detail structure of data
}
\examples{
myttest(x=rnorm(30,10,12), y=rnorm(40, 7, 10))
}
