% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{mynewf1}
\alias{mynewf1}
\title{Normal Distribution histogram of Random Variable X data}
\usage{
mynewf1(x)
}
\arguments{
\item{x}{vector of data}
}
\value{
histogram plot
}
\description{
This is a visualization of the distribution of the data set for random continuous variable X
}
\details{
The constructor function creates a histogram of the data set X.
}
\examples{
mynewf1(x=rnorm(100))
}
